selfiescan/requirements.txt
absl-py==2.3.1
albucore==0.0.23
albumentations==2.0.5
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
astunparse==1.6.3
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
beautifulsoup4==4.12.3
billiard==4.2.1
blinker==1.9.0
boto3==1.39.8
botocore==1.39.8
celery==5.4.0
certifi==2024.12.14
cffi==1.17.1
channels==4.2.2
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
coloredlogs==15.0.1
constantly==23.10.4
contourpy==1.3.1
cryptography==44.0.0
cycler==0.12.1
Cython==3.0.12
daphne==4.2.0
Django==5.1.4
django-allauth==65.3.1
django-environ==0.12.0
django-ratelimit==4.1.0
django-storages==1.14.6
djangorestframework==3.16.0
dlib==19.24.6
easydict==1.13
face-recognition==1.3.0
face_recognition_models==0.3.0
filelock==3.16.1
fire==0.7.0
Flask==3.1.0
Flask-Cors==5.0.0
flatbuffers==24.3.25
flower==2.0.1
fonttools==4.57.0
gast==0.6.0
gdown==5.2.0
google-pasta==0.2.0
grpcio==1.68.1
gunicorn==23.0.0
h11==0.16.0
h5py==3.12.1
httpcore==1.0.9
httpx==0.28.1
humanfriendly==10.0
humanize==4.12.2
hyperlink==21.0.0
idna==3.10
imageio==2.37.0
incremental==24.7.2
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
keras==3.7.0
kiwisolver==1.4.8
kombu==5.5.1
lazy_loader==0.4
libclang==18.1.1
lz4==4.3.3
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mdurl==0.1.2
ml-dtypes==0.4.1
mpmath==1.3.0
mtcnn==1.0.0
namex==0.0.8
networkx==3.4.2
numpy==2.0.2
oauthlib==3.2.2
opencv-python==4.10.0.84
opencv-python-headless==4.11.0.86
opt_einsum==3.4.0
optree==0.13.1
packaging==24.2
pandas==2.2.3
pillow==12.2.0
pillow-avif-plugin==1.5.5
pillow_heif==1.3.0
prettytable==3.16.0
prometheus_client==0.21.1
prompt_toolkit==3.0.50
protobuf==5.29.1
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
Pygments==2.18.0
PyJWT==2.10.1
pyOpenSSL==25.1.0
pyparsing==3.2.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
qrcode==8.0
razorpay==1.4.2
redis==5.2.1
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.4
s3transfer==0.13.0
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.2
service-identity==24.2.0
setuptools==75.6.0
simsimd==6.2.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
sqlparse==0.5.3
stringzilla==3.12.4
sympy==1.13.3
termcolor==2.5.0
threadpoolctl==3.6.0
tifffile==2025.3.30
tornado==6.4.2
tqdm==4.67.1
Twisted==25.5.0
txaio==25.6.1
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.0
zope.interface==7.2
